copy/paste format results code, will not refactor for now
