+++ _output/scans/1.11.1/gloo.txt
@@ -1,0 +1,15 @@
+
+
+Vulnerabilities Listed for quay.io/solo-io/gloo:1.11.1 (alpine 3.15.0)
+
+Vulnerability ID|Package|Severity|Installed Version|Fixed Version|Reference
+---|---|---|---|---|---
+CVE-2022-37434|zlib|CRITICAL|1.2.12-r0|1.2.12-r2|https://avd.aquasec.com/nvd/cve-2022-37434
+
+Vulnerabilities Listed for usr/local/bin/gloo
+
+Vulnerability ID|Package|Severity|Installed Version|Fixed Version|Reference
+---|---|---|---|---|---
+CVE-2022-21698|github.com/prometheus/client_golang|HIGH|v1.11.0|1.11.1|https://avd.aquasec.com/nvd/cve-2022-21698
+CVE-2022-27191|golang.org/x/crypto|HIGH|v0.0.0-20210920023735-84f357641f63|0.0.0-20220314234659-1baeb1ce4c0b|https://avd.aquasec.com/nvd/cve-2022-27191
+CVE-2021-44716|golang.org/x/net|HIGH|v0.0.0-20211205041911-012df41ee64c|0.0.0-20211209124913-491a49abca63|https://avd.aquasec.com/nvd/cve-2021-44716
