+++ securityscanutils/cmd/main.go
@@ -1,14 +1,15 @@
 package main
 
 import (
-    "context"
-    "github.com/solo-io/go-utils/securityscanutils/internal/commands"
-    "log"
+	"context"
+	"log"
+
+	"github.com/solo-io/go-utils/securityscanutils/internal/commands"
 )
 
 func main() {
-    ctx := context.Background()
-    if err := commands.RootCommand(ctx).Execute(); err != nil {
-        log.Fatalf("unable to run: %v\n", err)
-    }
+	ctx := context.Background()
+	if err := commands.RootCommand(ctx).Execute(); err != nil {
+		log.Fatalf("unable to run: %v\n", err)
+	}
 }
