+++ securityscanutils/constants.go
@@ -2,9 +2,6 @@ package securityscanutils
 
 const (
 	GithubRepositoryOwner = "solo-io"
-
-	QuayRepository = "quay.io/solo-io"
-
-	OutputScanDirectory = "_output/scans"
+	QuayRepository        = "quay.io/solo-io"
+	OutputScanDirectory   = "_output/scans"
 )
-
