+++ securityscanutils/internal/commands/scan_repo.go
@@ -1,90 +1,90 @@
 package commands
 
 import (
-    "context"
-    "fmt"
-    "github.com/Masterminds/semver/v3"
-    "github.com/solo-io/go-utils/cliutils"
-    "github.com/solo-io/go-utils/securityscanutils"
-    "github.com/solo-io/go-utils/securityscanutils/internal"
-    "github.com/spf13/cobra"
-    "github.com/spf13/pflag"
+	"context"
+	"fmt"
+
+	"github.com/Masterminds/semver/v3"
+	"github.com/solo-io/go-utils/cliutils"
+	"github.com/solo-io/go-utils/securityscanutils"
+	"github.com/solo-io/go-utils/securityscanutils/internal"
+	"github.com/spf13/cobra"
+	"github.com/spf13/pflag"
 )
 
 func ScanRepoCommand(ctx context.Context, globalFlags *internal.GlobalFlags) *cobra.Command {
-    opts := &scanRepoOptions{
-        GlobalFlags: globalFlags,
-    }
+	opts := &scanRepoOptions{
+		GlobalFlags: globalFlags,
+	}
 
-    cmd := &cobra.Command{
-        Use:   "scan-repo",
-        Short: "Scan a set of images/versions produced from a single github repository",
-        Long:  "Runs Trivy scans (only reports HIGH and CRITICAL-level vulnerabilities) against images from the repo specified and upload scan results to google cloud bucket",
-        RunE: func(cmd *cobra.Command, args []string) error {
-            return doScanRepo(ctx, opts)
-        },
-    }
-    opts.addToFlags(cmd.Flags())
-    cmd.SilenceUsage = true
-    return cmd
+	cmd := &cobra.Command{
+		Use:   "scan-repo",
+		Short: "Scan a set of images/versions produced from a single github repository",
+		Long:  "Runs Trivy scans (only reports HIGH and CRITICAL-level vulnerabilities) against images from the repo specified and upload scan results to google cloud bucket",
+		RunE: func(cmd *cobra.Command, args []string) error {
+			return doScanRepo(ctx, opts)
+		},
+	}
+	opts.addToFlags(cmd.Flags())
+
+	return cmd
 }
 
 type scanRepoOptions struct {
-    *internal.GlobalFlags
+	*internal.GlobalFlags
 
-    githubRepository string
-    imageRepository string
+	githubRepository string
+	imageRepository  string
 
-    // action to take when a vulnerability is discovered. supported actions are:
-    //  none (default): do nothing when a vulnerability is discovered
-    //  github-issue-latest (preferred): create a github issue only for the latest patch version of each minor version, when a vulnerability is discovered
-    //  github-issue-all: create a github issue for every version where a vulnerability is discovered
-    vulnerabilityAction string
+	// action to take when a vulnerability is discovered. supported actions are:
+	//  none (default): do nothing when a vulnerability is discovered
+	//  github-issue-latest (preferred): create a github issue only for the latest patch version of each minor version, when a vulnerability is discovered
+	//  github-issue-all: create a github issue for every version where a vulnerability is discovered
+	vulnerabilityAction string
 
-    releaseVersionConstraint string
-    imagesVersionConstraintFile string
+	releaseVersionConstraint    string
+	imagesVersionConstraintFile string
 }
 
 func (m *scanRepoOptions) addToFlags(flags *pflag.FlagSet) {
-    flags.StringVarP(&m.githubRepository, "github-repo", "g", "", "github repository to scan")
-    flags.StringVarP(&m.imageRepository, "image-repo", "r", securityscanutils.QuayRepository, "image repository to scan")
+	flags.StringVarP(&m.githubRepository, "github-repo", "g", "", "github repository to scan")
+	flags.StringVarP(&m.imageRepository, "image-repo", "r", securityscanutils.QuayRepository, "image repository to scan")
 
-    flags.StringVarP(&m.vulnerabilityAction, "vulnerability-action",  "a", "none", "action to take when a vulnerability is discovered (none, github-issue-all, github-issue-latest")
+	flags.StringVarP(&m.vulnerabilityAction, "vulnerability-action", "a", "none", "action to take when a vulnerability is discovered {none, github-issue-all, github-issue-latest}")
 
-    flags.StringVarP(&m.releaseVersionConstraint, "release-constraint", "c", "", "version constraint for releases to scan")
-    flags.StringVarP(&m.imagesVersionConstraintFile, "image-constraint-file", "i", "", "name of file with mapping of version to images")
+	flags.StringVarP(&m.releaseVersionConstraint, "release-constraint", "c", "", "version constraint for releases to scan")
+	flags.StringVarP(&m.imagesVersionConstraintFile, "image-constraint-file", "i", "", "name of file with mapping of version to images")
 
-    cliutils.MustMarkFlagRequired(flags, "github-repo")
-    cliutils.MustMarkFlagRequired(flags, "release-constraint")
-    cliutils.MustMarkFlagRequired(flags, "image-constraint-file")
+	cliutils.MustMarkFlagRequired(flags, "github-repo")
+	cliutils.MustMarkFlagRequired(flags, "release-constraint")
+	cliutils.MustMarkFlagRequired(flags, "image-constraint-file")
 }
 
-
 func doScanRepo(ctx context.Context, opts *scanRepoOptions) error {
-    releaseVersionConstraint, err := semver.NewConstraint(fmt.Sprintf("%s", opts.releaseVersionConstraint))
-    if err != nil {
-        return err
-    }
-    imagesPerVersion, err := internal.GetImagesPerVersionFromFile(opts.imagesVersionConstraintFile)
-    if err != nil {
-        return err
-    }
+	releaseVersionConstraint, err := semver.NewConstraint(fmt.Sprintf("%s", opts.releaseVersionConstraint))
+	if err != nil {
+		return err
+	}
+	imagesPerVersion, err := internal.GetImagesPerVersionFromFile(opts.imagesVersionConstraintFile)
+	if err != nil {
+		return err
+	}
 
-    scanner := &securityscanutils.SecurityScanner{
-        Repos: []*securityscanutils.SecurityScanRepo{
-            {
-                Repo:  opts.githubRepository,
-                Owner: securityscanutils.GithubRepositoryOwner,
-                Opts: &securityscanutils.SecurityScanOpts{
-                    OutputDir:                   securityscanutils.OutputScanDirectory,
-                    ImagesPerVersion:            imagesPerVersion,
-                    VersionConstraint:           releaseVersionConstraint,
-                    ImageRepo:                   opts.imageRepository,
-                    CreateGithubIssuePerVersion: opts.vulnerabilityAction == "github-issue-all",
-                    CreateGithubIssueForLatestPatchVersion: opts.vulnerabilityAction == "github-issue-latest",
-                },
-            },
-        },
-    }
-    return scanner.GenerateSecurityScans(ctx)
+	scanner := &securityscanutils.SecurityScanner{
+		Repos: []*securityscanutils.SecurityScanRepo{
+			{
+				Repo:  opts.githubRepository,
+				Owner: securityscanutils.GithubRepositoryOwner,
+				Opts: &securityscanutils.SecurityScanOpts{
+					OutputDir:                              securityscanutils.OutputScanDirectory,
+					ImagesPerVersion:                       imagesPerVersion,
+					VersionConstraint:                      releaseVersionConstraint,
+					ImageRepo:                              opts.imageRepository,
+					CreateGithubIssuePerVersion:            opts.vulnerabilityAction == "github-issue-all",
+					CreateGithubIssueForLatestPatchVersion: opts.vulnerabilityAction == "github-issue-latest",
+				},
+			},
+		},
+	}
+	return scanner.GenerateSecurityScans(ctx)
 }
